Pack the typescript types in the npm tarball
